==> src/edai_0001/04-images/02_explore_data.ipynb <==
# Explorar nosso Dataset

import pandas as pd

pd.options.display.max_columns = None
pd.options.display.max_rows = None
pd.options.display.max_colwidth = None


DATASET_FRACTION = 0.1

dataset_path = 'data/images/home/sdf/marketing_sample_for_amazon_com-ecommerce__20200101_20200131__10k_data.csv'

dataset_df = pd.read_csv(dataset_path).sample(frac=DATASET_FRACTION)
dataset_df.columns

#dataset_df.head(5)
dataset_df[['Product Name', 'Image']].head(5)

# Removendo última imagem (transparent pixel)
dataset_df["Image"] = dataset_df["Image"].map(lambda x: x.split("|")[:-1])

dataset_df[['Product Name', 'Image']].head(5)

# Explodindo a lista de imagens (uma linha para cada imagem)
dataset_df = dataset_df.explode("Image").dropna(subset=["Image"])
dataset_df[['Product Name', 'Image']].head(10)
